Cleans up some naming
